test-vector-generator/input-generators/requirements.txt
# Python dependencies for test vector generation
numpy==1.24.3
pyyaml==6.0
